[refactor]MangaCardをcommonへ移動
