+++ front/src/components/common/MangaCard.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
-    <v-card height="120px" width="80px" elevation="5">
-      <v-img :src="book.src" height="120px" width="80px"> </v-img>
+    <v-card :height="height" :width="width" elevation="5">
+      <v-img :src="book.src" :height="height" :width="width" />
     </v-card>
   </div>
 </template>
@@ -11,7 +11,7 @@ import Vue, { PropType } from "vue";
 import { Book } from "../../types/Book";
 
 export default Vue.extend({
-  name: "Home",
+  name: "MangaCard",
   props: {
     book: {
       type: Object as PropType<Book>,
@@ -20,6 +20,16 @@ export default Vue.extend({
         src: "https://1.bp.blogspot.com/-tVeC6En4e_E/X96mhDTzJNI/AAAAAAABdBo/jlD_jvZvMuk3qUcNjA_XORrA4w3lhPkdQCNcBGAsYHQ/s1048/onepiece01_luffy.png",
       }),
       required: false,
+    },
+    width: {
+      type: Number,
+      default: 0,
+      required: true,
+    },
+    height: {
+      type: Number,
+      default: 0,
+      required: true,
     },
   },
 });
