+++ front/src/components/mangaList/MangaListView.vue
@@ -2,7 +2,7 @@
   <div class="ma-5">
     <v-row v-for="n in rowNum" :key="n">
       <v-col v-for="m in colNum(n)" :key="m">
-        <MangaCard :book="books[(n - 1) * 3 - m]" />
+        <MangaCard :book="books[(n - 1) * 3 - m]" :width="80" :height="120" />
       </v-col>
     </v-row>
   </div>
@@ -10,7 +10,7 @@
 
 <script lang="ts">
 import Vue, { PropType } from "vue";
-import MangaCard from "./MangaCard.vue";
+import MangaCard from "../common/MangaCard.vue";
 import { Book } from "../../types/Book";
 
 export default Vue.extend({
